#CLOUD-1030 cloud_db: resetting sql query statistics after initial cache filling
#CLOUD-990 Added inactivity timeout.
#CLOUD-990 Fixed compile error.
#VMS-5461 Backport from 3.0.0
#VMS-5921 Added missing #include
#VMS-5921 Fixed compile errors
#VMS-5921 Renaming: <utils/common/subscription.h> -> <nx/utils/subscription.h>
#VMS-5921 smartSplit function moved to <nx/utils/string.h>
#VMS-5921 unquoteStr function moved to <nx/utils/string.h>
Always use release provisioning profile and entitlements to be able to use TestFlight
Backed out changeset: cc5cd9177d37
Closed
Compile fix after merge with vms_2.6
Fixed round error in unit test
Fixed ubjson double deserialization
MOBILE-466: Do not move timeline from top to bottom on opening camera from search results
MOBILE-508: Leave full-screen mode when exiting video screen
MOBILE-509: Display Video Screen controls if the app was re-activated
MOBILE-510: Animate timeline when first time tapping archive on offline camera
MOBILE-515: Fix Lite Client connection if server credentials were changed
MOBILE-528: Do not save cloud connection credentials
MOBILE-528: Remove last usages of firstHost
MOBILE-530: Elide text on check boxes
MOBILE-531: No archive on Samsung after merge
MOBILE: Fix invalid merge
MOBILE: Fix video item positioning at 1/3 of the screen height
MS-4748 - Press 'Cancel' in System administration dialog during update behavior
MS-5510 - Auto-login to cloud systems
MS-5954 - Mac Client. Small modal dialogs should not be hidden by parent dialogs
Removed some unused code
TEST-180 new proxy test, style fixes & improvements
VMS-5599 - Wrong number of required licenses
VMS-5885: Intermittent selection on 'Layout' tab in User Settings
VMS-5936: HardwareID field is empty at the License Tab of System Administration(DW customer's site)
VMS-5945 - Copying schedule doesn't work on mac
VMS-5948: Fast scan shouldn't pause if disk is slow
VMS-5949 - All buttons disappear from Video Wall item if user enabled info
VMS-5950: Apidoc for /api/getEvents incorrectly describes its result
VMS-5956 - Temporary login on server tile on Welcome screen
VMS-5957 - Disable 'Recording' form if recording is disable on camera
VMS-5960: Crash detected in server: 2be7a6e392ad00010768c0888f6979bd
VMS-5963:New dahua camera authentication fix
customizations
use nxcloud 1.0.134
vms-5862. nx1 UT optimizations build fix.
vms-5862. nx1 UT optimizations.

<==== Mercurial data ====>
Branch: vms_3.1_gui

+++ common_libs/nx_fusion/src/nx/fusion/serialization/ubjson_detail.h
@@ -31,7 +31,7 @@ namespace QnUbjsonDetail {
         return tmp.f;
     }
 
-    inline float fromBigEndian(double value) {
+    inline double fromBigEndian(double value) {
         /* Avoid breaking strict aliasing rules by using a union. */
         union {
             quint64 i;
